chore: extract getAiToken to service
